refactor: Create repository package and move all repositories
