+++ src/main/java/com/codessquad/qna/domain/repository/AnswerRepository.java
@@ -1,5 +1,6 @@
-package com.codessquad.qna.domain;
+package com.codessquad.qna.domain.repository;
 
+import com.codessquad.qna.domain.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AnswerRepository extends JpaRepository<Answer, Long> {
