+++ src/main/java/com/codessquad/qna/domain/repository/QuestionRepository.java
@@ -1,5 +1,6 @@
-package com.codessquad.qna.domain;
+package com.codessquad.qna.domain.repository;
 
+import com.codessquad.qna.domain.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
